+++ docs/source/changelog.rst
@@ -11,6 +11,13 @@ Changelog
 History
 ==========================
 
+Version 0.3.1 (2023-04-02)
+--------------------------
+
+* Fixed issue for path based query, where Rickle objects are considered callable (rightfully).
+* Uses `inspect.isfunction` instead of `callable`.
+* Added `meta` to base class for getting metadata of a property.
+
 Version 0.3.0 (2023-02-16)
 --------------------------